Complete assignment 3
Learned that I tend to add additional code:
In reviewing the solution, my visual did match the desired outcome image
but comparing the solution.css files, I had a few extra lines including
padding, or trying to align the child items.  In this instance, they did
not interfere with the final image or functionality, but I imagine in
future they could.

The solution might be to use gap, rather than padding.